// packages/plugins/MagicScroll/tests/enums.ts
// ─── TestId ───────────────────────────────────────────────────────────────────
// Merged from all MagicScroll test files

export enum TestId {
  // rendering.test.ts
  HasScroll = 'has-scroll',
  Progress = 'progress',
  // options.test.ts
  Scroller = 'scroller',
  ProgressA = 'progress-a',
  ProgressB = 'progress-b',
  // collision.test.ts
  Scroller2 = 'scroller2',
}